lock: Avoid remove incoherent exclusive roster
If the exclusive lock dir is not present, but the roster contains an
exclusive lock, in the process of unwinding the panic, the code would
remove the roster file.

Handle this by explicitly checking for this situation, and removing the
in-memory roster so that the file will be left alone.